web: refactored code in core and web components Signed-off-by: Muhammad Ali <alihamuh@gmail.com>
