+++ apps/web/src/components/dialogs/move-topics-dialog.tsx
@@ -17,21 +17,16 @@ You should have received a copy of the GNU General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-import { ChangeEvent, useCallback, useEffect, useRef, useState } from "react";
-import { Button, Flex, Text } from "@theme-ui/components";
+import { useCallback, useState } from "react";
+import { Flex, Text } from "@theme-ui/components";
 import * as Icon from "../icons";
 import { db } from "../../common/db";
 import Dialog from "./dialog";
-import Field from "../field";
 import { useStore, store } from "../../stores/notebook-store";
 import { Perform } from "../../common/dialog-controller";
+import FilteredList, { Item } from "../filtered-list";
 
 type MoveDialogProps = { onClose: Perform; topics: []; id: String };
-type Item = {
-  id: string;
-  type: "notebook" | "header";
-  title: string;
-};
 type Notebook = Item;
 
 function MoveDialog({ onClose, topics, id }: MoveDialogProps) {
@@ -123,107 +118,4 @@ function MoveDialog({ onClose, topics, id }: MoveDialogProps) {
   );
 }
 
-type FilteredListProps<T extends Item> = {
-  placeholders: { filter: string; empty: string };
-  items: () => T[];
-  filter: (items: T[], query: string) => T[];
-  onCreateNewItem: (title: string) => Promise<void>;
-  renderItem: (item: T, index: number, refresh: () => void) => JSX.Element;
-};
-
-function FilteredList<T extends Item>(props: FilteredListProps<T>) {
-  const {
-    items: _items,
-    filter,
-    onCreateNewItem,
-    placeholders,
-    renderItem
-  } = props;
-
-  const [items, setItems] = useState<T[]>([]);
-  const [query, setQuery] = useState<string>();
-  const noItemsFound = items.length <= 0 && query && query.length > 0;
-  const inputRef = useRef<HTMLInputElement>(null);
-
-  const refresh = useCallback(() => {
-    setItems(_items());
-  }, [_items]);
-
-  useEffect(() => {
-    refresh();
-  }, [refresh]);
-
-  const _filter = useCallback(
-    (query) => {
-      setItems(() => {
-        const items = _items();
-        if (!query) {
-          return items;
-        }
-        return filter(items, query);
-      });
-      setQuery(query);
-    },
-    [_items, filter]
-  );
-
-  const _createNewItem = useCallback(
-    async (title) => {
-      await onCreateNewItem(title);
-      refresh();
-      setQuery(undefined);
-      if (inputRef.current) inputRef.current.value = "";
-    },
-    [inputRef, refresh, onCreateNewItem]
-  );
-
-  return (
-    <>
-      <Field
-        inputRef={inputRef}
-        data-test-id={"filter-input"}
-        autoFocus
-        placeholder={
-          items.length <= 0 ? placeholders.empty : placeholders.filter
-        }
-        onChange={(e: ChangeEvent) =>
-          _filter((e.target as HTMLInputElement).value)
-        }
-        onKeyUp={async (e: KeyboardEvent) => {
-          if (e.key === "Enter" && noItemsFound) {
-            await _createNewItem(query);
-          }
-        }}
-        action={
-          items.length <= 0
-            ? {
-                icon: Icon.Plus,
-                onClick: async () => await _createNewItem(query)
-              }
-            : { icon: Icon.Search, onClick: () => _filter(query) }
-        }
-      />
-      <Flex mt={1} sx={{ overflowY: "hidden", flexDirection: "column" }}>
-        {noItemsFound && (
-          <Button
-            variant={"secondary"}
-            sx={{
-              display: "flex",
-              justifyContent: "space-between",
-              alignItems: "center",
-              py: 2
-            }}
-            onClick={async () => {
-              await _createNewItem(query);
-            }}
-          >
-            <Text variant={"body"}>{`Add "${query}"`}</Text>
-            <Icon.Plus size={16} color="primary" />
-          </Button>
-        )}
-        {items.map((item, index) => renderItem(item, index, refresh))}
-      </Flex>
-    </>
-  );
-}
 export default MoveDialog;
