+++ apps/web/src/components/topic/index.js
@@ -93,7 +93,6 @@ const menuItems = [
     title: "Move Topic",
     icon: Icon.Move,
     onClick: ({ items, notebookId }) => {
-      console.log(items);
       showMoveTopicsDialog(items, notebookId);
     },
     multiSelect: true
